add -s, --single-line option for single line output
To use this output with the GitHub API to create a new release, the
commit message needs to include literal \n characters, e.g.
"-change 1\n- change 2\n- change 3" instead of passing the actual
newline characters themselves.

This change allows for generation of a single line of output instead of
the default multi-line output.